requirements.txt
celery
scandir
pafy
django >=1.8, <1.9
djangorestframework >= 3.1.0
django-cors-headers>=1.0.0